+++ R/PipeOpTaskSurvRegr.R
@@ -279,6 +279,7 @@ PipeOpTaskSurvRegr = R6Class("PipeOpTaskSurvRegr",
 
       x = data.frame(backend)[, colnames(backend) %nin% c(time, status), drop = FALSE]
       x = model.matrix(~., x)[, -1L]
+
       bj = invoke(bujar::bujar,
         y = backend[[time]],
         cens = backend[[status]],